pyproject.toml
[project]
name = "diffsinger-voicevox-connector"
version = "0.1.0"
description = "VOICEVOX Engine API-compatible HTTP server that routes singing synthesis to DiffSinger (openvpi) ONNX models"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = { text = "MIT" }
authors = [{ name = "uboar" }]

dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "onnxruntime>=1.18,<2.0",
    "numpy>=1.26,<2.1",
    "scipy>=1.11,<2.0",
    "librosa>=0.10,<1.0",
    "soundfile>=0.12,<1.0",
    "pyopenjtalk>=0.3.4",
    "PyYAML>=6.0,<7.0",
    "click>=8.1,<9.0",
]

[project.optional-dependencies]
gpu = ["onnxruntime-gpu>=1.18,<2.0"]
dev = [
    "pytest>=8.2,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "httpx>=0.27,<1.0",
    "ruff>=0.5,<1.0",
]
build = [
    "pyinstaller>=6.6,<7.0",
]

[project.scripts]
diffsinger-connector = "diffsinger_engine.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/diffsinger_engine"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
